fix: More defensive code when trying to drop tables

+++ migration/2_ansel_upgrade_autoincrement.php
@@ -18,24 +18,20 @@ class AnselUpgradeAutoIncrement extends Horde_Db_Migration_Base
     public function up()
     {
         $this->changeColumn('ansel_images', 'image_id', 'autoincrementKey');
-        try {
+        if (in_array('ansel_images_seq', $this->tables())) {
             $this->dropTable('ansel_images_seq');
-        } catch (Horde_Db_Exception $e) {
         }
         $this->changeColumn('ansel_faces', 'face_id', 'autoincrementKey');
-        try {
+        if (in_array('ansel_faces_seq', $this->tables())) {
             $this->dropTable('ansel_faces_seq');
-        } catch (Horde_Db_Exception $e) {
         }
         $this->changeColumn('ansel_shares', 'share_id', 'autoincrementKey');
-        try {
+        if (in_array('ansel_shares_seq', $this->tables())) {
             $this->dropTable('ansel_shares_seq');
-        } catch (Horde_Db_Exception $e) {
         }
         $this->changeColumn('ansel_tags', 'tag_id', 'autoincrementKey');
-        try {
+        if (in_array('ansel_tags_seq', $this->tables())) {
             $this->dropTable('ansel_tags_seq');
-        } catch (Horde_Db_Exception $e) {
         }
     }
